Implemented backward animation on patient screen.
Currently, the main screen is not updated on AppStateRequiredActionData events (which is only a problem for delayed events such as uploads to VisibleImpact not working). The reason why we do not update the requiredActions on the main screen is because it interferes with the backward animation on the patient screen, i.e., the setState in the main screen triggers a re-render of the patient screen before the backward animation has completed.